Added GameScreen class
GameScreen.cs
Acts as a layout for the board and other objects which will appear on
the game screen.

Board.cs
Added handlers to pass legal Move objects to GameScreen. These will be
used to add text to the side text box things.

Game.xaml.cs
Edited to now create a GameScreen instead of a Board.